Resolve btc address based on btcutil encoding
An address provided as a beneficiary address can match one of many bitcoin
addresses formats. One of which is P2PK where a public key is provided.
Previously we would return the same public key that was suplied by the
user, but btcutil.DecodeAddress resolves a correct address for this
public key, so we want to use the address resolved by btcutil.

Here we updated unit tests to cover the scenario described above
but also enhanced the tests a little bit with some test cases taken from
https://github.com/btcsuite/btcutil/blob/e2ba6805a89060647fa22ec61b799acff6a4f6fd/address_test.go