Validate Bitcoin addresses received from peer members
We want to validate Bitcoin addresses received from peer members on
liquidation recovery to confirm the addresses are valid for the network
the client is supporting. Without this validation the client may accept
an address from a missconsifured peer which was desired for a different
network.

The validation here differs from `ValidateAddress` defined in
`derive_address.go` as here we don't expect to receive *pub keys from
peers, as this is peer's duty to derive a bitcoin address from *pub key
before broadcasting it.