Initial tests for liquidation recovery logic
We want to test logic for liquidation recovery. The tests are focused on
testing logic with different states of bitcoin connection. It's expected
that if bitcoin connection fails the process moves on and completes
based on local data.

For the time being the signers are generated with TSS protocol but it
would be awesome to load them from test data. Unfortunatelly I was not
able to get it working in the limited time frame so leaving it as is.