Align timeouts for electrs connection and liquidation recovery protocol
A timeout for electrs connection retries has to be less than the timeout
used for the liquidation protocol execution, so the nodes can correctly
sychronize.

Currently electrs connection timeout is the same as default retry timeout
for liquidation recovery (2 minutes). So it’s hard for the nodes to sync
in the same time frame to execute the complete protocol. They all announce
presence so node 1 which has non-working electrs connection retries for 2 minutes,
and once it aborts electrs retries and proceed to next protocol step the other
nodes hit the 2 minutes timeout waiting for the node 1 and start again
from the begining.