Removed the deployment considerations section
This section points to an example of running multiple ecdsa nodes in a
kube cluster. While this is useful, I feel that it dilutes the message
and intent of the document, which is to get the user started running a
node. If we want to keep this information, I think we should find a
different home for it.