Add a workaround for a problem with `prettier-config-keep` dependency
Since recently we have two package-lock.json configs - one in the
`keep-ecdsa` and one in `keep-ecdsa/solidity`.
The `@keep-network/prettier-config-keep` dependency is defined in the
former one, and is not present in the latter one. When `npm run lint` is
run in the `keep-ecdsa/solidity`, the manager looks not only at the
config file in that folder, but also at the config in the root directory.
It throws an error about missing `@keep-network/prettier-config-keep` if
`npm ci` isn't run in the root directory.
Due to that we're adding a workaround that installs the dependencies not
only in the `solidity` folder, but also in the project's root directory.