Improved liquidation errors bubbling up and logging
As we refactored liquidation recovery handling to be extracted to a
separate function we can tweak logging too.

In BuildBitcoinTransaction `keep` handle is now used just to put the
keep ID in the error message so we get rid of it and expect the ID to be
added by the parent funcion. Additionally we no longer log errors but
wrap them and bubble up to `handleLiquidationRecovery` caller.