Fix conditions for execution of jobs publishing the image
The conditions for executing `client-build-test-publish-ethereum` and
`client-build-test-publish-celo` jobs were not set up properly,
resulting in execution of both jobs when triggered by the
`workflow_dispath` event (no matter what was the value of `environment`
input).
This commit fixes the conditions, so that only one job of the two would
be executed when workflow gets triggered by the `workflow_dispatch'.