CLI command to resolve bitcoin beneficiary address
Operator configures the bitcoin beneficiary address in the config file.
The value can be *pub extended key that is used to derive a next
available unused bitcoin address. We want to give the operator a way to
verify client configuration resolves to an address that they desire to
use.

By running a `keep-ecdsa resolve-bitcoin-address` command the same logic
for address resolution that for liquidation recovery will be used. The
printed address will be the address that would be used if the
liquidation recovery took place.

The address resolution will be performed in dry-run, which means the
resolved address index won't get stored on the disk storage. So when
liquidation recovery is run it will use the same address.

It also gives the operator a way to verify bitcoin chain handle
connection as resolve protocol calls bitcoin chain to verify address
usage.