Move initcontainer job out of Solidity workflow to a separate workflow
In separate commits we're changing the kube configuration for
`keep-ecdsa` pods, so that the population of `TBTCSystem` address would
happen automatically. For the feature to work correctly we need to build
the `keep-ecdsa` initcontainer after migrating `tbtc` contracts. To do
that, we need to have separate workflows for `keep-ecdsa` contracts
migration and for initcontainer building/publishing, which we're doing
in this commit.
Apart from that, we'll be changing the CI flow configuration in the
`keep-network/ci/config/config.json` which orchestrates the order of
workflows execution during deployment on testnet.