Abort liquidation recovery for not funded deposit
A keep can get terminated before a deposit has been funded. This happens
in a situation when `notifySignerSetupFailed` is called for the deposit.
Currently the client will exit with panic as the funding info is not
available. With the change in this commit the client will check if the
deposit has been funded, if not it will exit liquidation recovery retry
loop without an error to abort retries.