Remove always() from if check
We used `always()` in keep-core where we wanted to execute the job
always, no matter of the result of jobs that it depended on. This is
used in keep-core to not depend on the result of solidity docs
generation.

Here we don't generate solidity docs yet so no need for this check.