Change timeout for WeiBalanceOf to 30 seconds
The request to `BalanceAt` should be quite simple and completed easily.
We decreased the timeout to 30 seconds as we noticed that from time to
time errors were thrown about `context deadline`, but 1 minute was too
much to hang the request. We have retries now for balance monitoring so
even if we're stuck for 30 seconds the subseqent retry should complete
the request.