Fix ignore comment for PasswordEnvVariable gosec
gosec command doesn't ignore the line if `// #nosec` is not in the first
comment line.
So what would work is:
```
// #nosec G101 (look for hardcoded credentials)
// This line doesn't contain any credentials.
// It's just the name of the environment variable.
// PasswordEnvVariable environment variable name for ethereum key password.
const PasswordEnvVariable = "KEEP_ETHEREUM_PASSWORD"
```
But this makes go lint fail with `exported const PasswordEnvVariable
should have comment or be unexported`.

The solution in this commit makes both golint and gosec work.